% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spoeff.R
\name{plot_spoeff}
\alias{plot_spoeff}
\title{Graph sporulation efficiency data}
\usage{
plot_spoeff(
  file_dir,
  replicates = c("TRUE", "FALSE"),
  treatment = c("TRUE", "FALSE")
)
}
\arguments{
\item{file_dir}{file directory to excel file with sporulation efficiency data with the columns strain, no, monads, dyads, tetrads, replicates (if applicable), and treatment (if applicable).}

\item{replicates}{"TRUE" or "FALSE", to denote whether the data has biological replicates}

\item{treatment}{"TRUE" or "FALSE", to denote whether the data has samples treated with a drug or another chemical}
}
\value{
a graph with \% of spores that formed tetrads on y axis, and strains on X axis
}
\description{
Graph sporulation efficiency data
}
