% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_viability.R
\name{plot_viability}
\alias{plot_viability}
\title{Graph spore viability data}
\usage{
plot_viability(
  file_dir,
  replicates = c("TRUE", "FALSE"),
  treatment = c("TRUE", "FALSE")
)
}
\arguments{
\item{file_dir}{file directory to excel file with spore viability with the columns strain, alive, dead, replicates (if applicable), and treatment (if applicable).}

\item{replicates}{"TRUE" or "FALSE", to denote whether the data has biological replicates}

\item{treatment}{"TRUE" or "FALSE", to denote whether the data has samples treated with a drug or another chemical}
}
\value{
a graph with \% of spores viable on y axis, and strains on X axis
}
\description{
Graph spore viability data
}
